Update install_dependencies.sh
It downloaded another file, but did not extract it, leading to an empty treetagger bin file. This should fix the issue and download the correct file from the treetagger server. After which their install_tagger.sh script should extract the folder. 

This is the part in the treetagger install_taggel.sh script:

if [ -r tree-tagger-linux-3.2-old.tar.gz ]
then
    gzip -cd tree-tagger-linux-3.2-old.tar.gz | tar -xf -
    echo 'Linux version of TreeTagger (older kernels) installed.'
fi

Therefor it does not extract tree-tagger-linux-3.2.tar.gz